requirements.txt
anndata==0.7.6 
bokeh==2.3.1
h5py==2.10.0
imageio==2.9.0   
leidenalg==0.8.4 
matplotlib==3.3.4 
numpy==1.20.2
python-igraph==0.9.1
python-louvain==0.15 
scanpy==1.7.2 
scikit-learn==0.24.1
scikit-network==0.23.1
scipy==1.6.3
seaborn==0.11.1
torch==1.8.1
torch-geometric==1.7.0 
torchvision==0.9.1
tqdm==4.59.0
umap-learn==0.5.1 